requirements.txt
snowflake-connector-python==3.6.0
snowflake-sqlalchemy==1.5.3

# Prometheus
prometheus-client==0.20.0

# OpenTelemetry
opentelemetry-sdk==1.24.0
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0

# AWS
boto3==1.34.0
botocore==1.34.0

# Apache Iceberg
pyiceberg[s3]==0.7.0
pyarrow==15.0.0

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.1
schedule==1.2.1
pandas==2.2.0
loguru==0.7.2
requests==2.31.0
